OSS: Configure kernel module, event formatter, and split extractor
We use a configuration that is passed as the value of an environment
variable at extractor startup. This configuration contains the set of
syscalls that we want to extract. We process these calls and pass them
along to the kernel as a module parameter. The kernel module is not
brought up on start, but is instead inserted programmatically after
we have processed the config and built the config arguments to the
kernel.

The event formatter is now configurable through the incoming config
file. A map that defines what to use for the sparse vector representation
is used to configure the formatter.

This commit also splits the extractor into two containers. One will
continue to be called extractor and will extract docker events and the
other called collector will extract sysdig events.

JIRA:
SROX-5194, SROX-5198, SROX-5486, SROX-5196