Fix: Kernel module insertion failures due to our attempt at "soft" removal
When the kernel module is already present, we attempt to remove it before
trying to re-insert it. We do this till the insertion succeeds. The problem
though is that we use a "soft" removal strategy that waits till the module
reference count goes down to zero. This may never happen if a prior run did
not cleanly exit, thus dropping the module reference count in the kernel.
Using O_TRUNC now to force module removal before re-trying insertion.

Fixed the text that is shown before kernel module insertion. Removed
the log messages in the insertion/deletion loop.

SROX-5776