Fix: Remove references to StackRox and extractor were not required
This change removes a few occurrences of extractor that were not required
Renames extractorContainerID to containerID

Removed `MAINTAINER StackRox` and also removed the print to stdout of the
M line. This enables removing GENERIC_SHORT_CONTAINER_ID, which is also
done. Removes references to our repositories from the test yaml files

With these changes the OSS epic is indeed complete

JIRA:
SROX-6409