SROX-8953: Update open-source attribution notice for 1.2
The open-source attribution notice is updated to match new deps.

CI will now trigger appropriate builds when NOTICE.txt changes,
and will trigger a roxc build when the notice.go file is updated.

The build trigger check is adjusted to match when the modified
filename exactly equals the build trigger, allowing NOTICE.txt
and notice.go to trigger separate builds.

The location of the Zookeeper NOTICE.txt is fixed to /NOTICE.txt.

The sysblock NOTICE is added to Collector, and the existing notice
is renamed to support both being present.