SROX-8820: Produce chisels dynamically and communicate them via Kafka
In Kubernetes 1.6 on Google Container Engine, container launch
configurations changed to use SHA256 Docker image IDs instead of
human-readable image tags (names).

To fix this, the image-exclusion chisel is now generated dynamically in
Director. Collectors listen for updates to the chisel on a topic called
"rox-chisels". Upon receiving a new chisel, the collector reinitializes
the sysdig consumer.

Rogue Container alerts consult the same whitelist, which has the effect
of suppressing such alerts for many orchestrator containers.

Various issues are corrected in the sysdig userspace program to enable
the shutdown and reinitialization of the consumer.

CI targets are tightened up a little to prevent rebuilding kernel
modules when only the userspace sysdig program changes.

The trainer Docker API version is fixed to support GKE (Docker 1.11).

A sleep is added between malicious training and detection pending SROX-9182.