SROX-10836: Upgrade to Bazel 0.8.1 in CI
Upgrade the build image to have Bazel 0.8.1, instead of the previous
0.5.4.

Update the Go plugin to 0.8.0 with a few more commits to fix the
--features=pure flag.

Fix the ML build to reference Protobuf jars the way that new Bazel wants
them to be referenced.

Bump bazel protoc and protogen.mk to 3.5.0.
Remove com_google_protobuf_java.
Upgrade collector protobuf to 3.5.0 to match all others.