SROX-9128: Don't use perpetually-out-of-date packages.ubuntu.com
packages.ubuntu.com is often not kept up to date until long after
kernels show up in mirrors and are used in clouds. Go straight to
the official mirrors instead of using the packages site.

Also, exclude kernel 4.13, which fails to compile in sysblock.
These kernels were not present in xenial-updates but are part of
the most recent non-LTS release.