SROX-12357 disable network health checks in global services
polling for network health status in the following services
is disabled when multi-cluster is on:
* collector
* container-summarizer
* enforcer

scanner-agent is already off in multi-cluster.

There's no reasonable method to collect that data from every service in
multi-cluster, and it's 99.9% redundant with having launcher already talk
to director. We do need to have launcher collect network reachability
tests for kafka, but that's coming separately.