ROX-3377: fixes eBPF probes on RHEL 7. (#535)
During normal operation, sysdig relies on the /proc file system to retrieve pid namespace information for processes that exist at the time that collector starts up (i.e. those that exist before we load the eBPF program and start collecting events) On Kernels newer than 4.1 this information is available in /proc/{pid}/status, under various NS* keys. However, on older kernels, this information does not exist.

If using kernel-module collection, the lack of procfs information is fine because the kernel module contains a custom ioctl that allows sysdig to query the kernel for the missing information directly, but this obviously does not work for eBPF collection, which does not have the ioctl installed. However, sysdig also attempts to identify containerized processes and sets a heuristic-based container ID where possible.

The fix is to combine these two pieces of information (vpid and container_id) to attempt to identify containerized processes. If we have a vpid, then the simple comparison against the thread's pid is enough, but if we don't have a vpid (like on RHEL 7) then we can attempt to use the container_id as the check.