ROX-8910 UEFI Secure Boot
When the system is boot in UEFI mode, the Secure Boot feature could be
enabled, preventing collector from loading unsigned kernel modules.
Recognize such situations and switch to BPF when needed. Generally
speaking there three options:

* the system is loaded in the legacy mode
* the system is loaded in UEFI mode with the SecureBoot disabled
* the system is loaded in UEFI mode with the SecureBoot enabled

The second option is tricky, because SecureBoot still could be enabled,
but the boot loaded didn't pass this information to the kernel. To
recornize the status of secure boot there are few options on the table
(e.g. read SecureBoot-* value from the efivarfs, or ask bootctl), but
the only way to check the undetermined state is to read boot_params
file. Since this also gives the rest of the information
(enabled/disabled), use only this approach for simplicity.

[1]: https://www.kernel.org/doc/html/latest/x86/zero-page.html