ROX-9462 Collect baselines (#617)
ROX-9462  Collect baselines

Introduce a script to manage performance baselines numbers. The baseline
is essentially a json file on the same GCS bucket as other CI results,
containing previous benchmark results. Its size is limited, the exact threshold
has to be figured out depending on the actual variance of data.

The script supposed to collect baseline, and provide way to compare current
benchmark data with it. The only metric we look at is `hackbench_avg_time`, but
baseline includes everything produced by the benchmark steps. Due to
unavoidable noise try to be fancy and employ t-test for mean values to figure
out if the difference we've got is significant, or just a noise.

This change also introduces a slight change in `all-perf.json` format,
before it was a list of json documents concatenated together, like:

    {"aaa": 1, "bbb": 2, ...} {"aaa": 3, "bbb": 4, ...}

This was fine for jq, but it's not a valid json. Not it will be:

    [{"aaa": 1, "bbb": 2, ...}, {"aaa": 3, "bbb": 4, ...}]

jq apparently didn't notice a thing.

Co-authored-by: Mauro Moltrasio <mmoltras@redhat.com>