Add context to be used on development builds (#656)
A new context has been added under 'collector/container/devel'.
The purpose of this context is to reuse the existing Dockerfile
for building collector images and enable installation of tools
such as valgrind and AddressSanitizer into it without any
possibility of said binaries ending up in a production image.

The new separation of contexts also makes it easy to add any
other tooling we might need in the future. Since the development
image is based on centos stream and dnf is not disabled, tools
can also be installed on the fly while debugging collector.