Have CI build all necessary QA containers (#674)
This commit makes it so all containers needed to
run integration tests and benchmarking are built
by our CI and pushed to quay.io periodically.

The containers will not be built in PRs if the
build-qa-containers label is not present. Layer
caching is enabled in the job in charge of
building the containers.