Bump Ubuntu VMs to 22.04 LTS in tests (#673)
Some extra changes on performance gathering:

Whenever a benchmark is run and a set of data is missing
from either the baseline or the current measurement,
ignore this instead of failing.

Running the update for the baseline will drop tests that
are not found in the new measurement.

There is no strict need to verify compatibility between baseline and the
new benchmark data, but some other assumption are still better to be
checked. E.g. presence of both benchmark runs with and without overheads
is necessary to make sure the data is meaningful.

On the way do some cosmetic adjustments: use more descriptive
"baseline_data" name for the argument of add_to_baseline_file, improve
commentary for process function, rename "process_data" to
"intersection".

Co-authored-by: Dmitrii Dolgov <9erthalion6@gmail.com>