Fix updating baseline data (#690)
Fix updating baseline data in 7432452 . The issue was about measure iterator,
which was used directly for producing the result. As soon as other operations
started to consume the iterator before that, it becomes empty, and the update
is not correctly performed.

On the way do cosmetic improvements for the installation process. Baseline
requires some dependencies, installed via pip. Do this silently to reduce
noise.