Use only root and the instance user from inside the container.
When running in rootless mode, podman will map both root and the
user running the instance (inside) to the user running the instance
(outside).
TAR, when run as root, tries to restore the owner/group store
in the archive, which then fails in podman.

We revert to use the usual behaviour of TAR (non-root) and exctact
the files as ourselves.