ROX-10009: custom eBPF probe (#670)
Adds a new eBPF probe front-end to the Falco plumbing which traces only
the specific tracepoints that collector is interested in. This drastically
improves system performance and allows much more targeted retrieval
of the required information.

Currently there exists within this probe a sys_enter/sys_exit approach
specifically for RHEL-7 which does not have the necessary tracepoints.