Integration tests retry
Add a simple retry function, apply it to the command executor and
container launcher. The implementation has few quirks:

* It doesn't use routines, but in case of integration tests it doesn't
matter if we would block.

* It has two return arguments with fixed type: string and error. The
latter one is required for retry logic, the former seem to be rather
limited choice. The reason is that without returning any data except an
error the scope of the payload function would have to be much bigger (as
nothing has to be returned). At the moment it looks like we mostly
return strings anyway, so it feels like a good compromise between
full-fledged solution and a quick hack.