ROX-12484: drive integration tests using Ansible (#871)
Adds the necessary roles and playbooks to run the integration tests
through Ansible. This includes creation of GCP VMs upon which to
run the tests, CI and Dev dynamic inventories as well as collating test results. 

Modifications to the integration tests themselves allow a 
one-to-many relationship between the test runner and hosts
under test to simplify CI.