Split builder and dev images
The purpose of this change is more organizational than anything else. We
currently build the builder image as a development environment, which
means we install quite a few packages in the image that don't actually
serve the purpose of building collector.

Splitting the image provides us with:
- More granularity and visibility into what packages are required
  for building collector and which for development.
- Reduced image size for the builder.

This first commit does not add the development image the our CI, instead
relies on the builder image built by CI as cache for building the
development image locally.

I've also take this opportunity to move the 'third_party' directory
inside the 'builder' directory, since it is only used when building the
builder image and allows for a smaller context to be sent over.