Add a script for building drivers and sinsp-example
The script is meant to be run from inside a collector-builder container,
so it should be portable enough to be run in multiple systems for
testing. We also add 'cmake' to the driver builder containers, since
this is a requirement to run the dev/build-drivers script.

We also remove the 'scripts' directory, since it had a single unused
script in it, I'd be happy to move this to a separate commit if needed.