Remove cmake from rhel7 builder
Version used for this is really old, we would need to enable the epel
repository and install the cmake3 package instead. Since I don't want to
accidentally install an epel dependencies and risk silently breaking the
drivers built with it, I think it is better to exclude it, hopefully no
one will need to run the sinsp-builder.sh script for older kernels.