ROX-13628 Create an external representation of endpoints (#978)
In Collector, an endpoint now represents a listening socket with a reference to the process making use of it. When an endpoint is published, however, the process information is stripped down to represent a the name and parameters of this process instead of the precise instance.

This as an impact on the life-cycle of endpoint objects "as seen from the outside". For instance, the same endpoint opened by several processes having the same attributes should appear as a single endpoint.

In order to implement this model, we change the comparison function used when fetching endpoints.
This will squash them during the normalization phase.

Co-authored-by: JoukoVirtanen <jouko@stackrox.com>