Add COLLECTOR_DRIVER variable
This environment variable is mostly meant for testing. It allows for
collector to look for a non-downloadable driver that should be added to
its filesystem (either by embedding it into the image or by bind
mounting it). If the provided path is not fully qualified, it is assumed
to be relative to '/kernel-modules'.