Simplify logging and unify levels in GetKernelObject
When downloadKernelDriver fails, a fatal error message is issued and
when a single driver fails in GetKernelObject a corresponding warning
log is issued. These two make the error message that was being printed
by downloadKernelDriver irrelevant and even hurts readability of logs in
my opinion.

As for the log levels in GetKernelObject, when it fails, a single driver failed
but we might be able to fallback to another driver, so I believe warnings are
more appropriate than errors.