Stacktrace improvements
Collector already uses a signal handler to report the stacktrace in case
of failures, but doesn't report all the symbols because:

* Collector is build without `-rdynamic`, missing some set of symbols in
the dynamic symbols table [1]. Adding it means the binary will become
bigger on the amount of added symbols (5.9 Mb vs 4.7 Mb without it).
Worth noting that those are not debugging symbols.

* Collector is build without `-fno-omit-frame-pointer`, potentially
missing some part of the stacktrace [2]. Adding it means a slight
performance overhead, but it's well known that it's negligible. E.g.
Fedora project recently switched to use it by default [3].

Adding those two options will give more readable stacktrace, that
contains reasonable symbolic names for functions. But the C++ parts are
going to be mangled and less readable. To address this, use `Dl_info`
and `abi::__cxa_demangle` [4] if possible.

Here is how the new stacktrace would look like:

    /lib64/libc.so.6 (null) 0x7f8b3df4bb80 + 0
    /usr/local/lib/libsinsp-wrapper.so check_api_compatibility 0x7f8b40c36373 + 12
    /usr/local/lib/libsinsp-wrapper.so (null) 0x7f8b40c344e9 + 0
    /usr/local/lib/libsinsp-wrapper.so scap_open_live_int 0x7f8b40c1be01 + 413
    /usr/local/lib/libsinsp-wrapper.so scap_open 0x7f8b40c1c969 + 279
    /usr/local/lib/libsinsp-wrapper.so sinsp::open_common(scap_open_args*) 0x7f8b40931a34 + 236
    /usr/local/lib/libsinsp-wrapper.so sinsp::open_bpf(std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, unsigned long, std::unordered_set<unsigned int, std::hash<unsigned int>, std::equal_to<unsigned int>, std::allocator
    collector collector::SysdigService::Start() 0x75467b + 1155
    collector collector::CollectorService::RunForever() 0x6ccc86 + 2396
    collector main 0x6bb0d8 + 1593
    /lib64/libc.so.6 __libc_start_main 0x7f8b3df37d85 + 229
    collector _start 0x6b916e + 46

[1]: https://gcc.gnu.org/onlinedocs/gcc/Link-Options.html
[2]: https://gcc.gnu.org/onlinedocs/gcc-4.9.2/gcc/Optimize-Options.html
[3]: https://fedoraproject.org/wiki/Changes/fno-omit-frame-pointer
[4]:
https://gcc.gnu.org/onlinedocs/libstdc++/libstdc++-html-USERS-4.3/a01696.html