Improve compatibility with podman
A small set of changes that enables using podman as the build engine via
the podman-docker package or by symlinking docker to podman.

The collector/falcosecurity-libs symlink is removed. It is not needed
now that we use a global cmake project set at the root of the repository
and was causing trouble with podman.

The '--cache-from' argument takes different types of arguments for
podman and docker, since the way the builder works makes it relatively
hard to actually take advantage of caching, we can directly remove them.