Auto cancel concurrent jobs in PRs
When working in PRs, we some times need to iterate fast and that means
pushing multiple times to a branch. This causes CI to be triggered
multiple times and we may need to cancel previous runs manually. This
small change makes it so previous CI jobs are cancelled when a new
commit is pushed to a PR (pushes to master and scheduled workflows
should be unaffected).

See https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#concurrency
for more information.