ebpf probe changes for s390x enablement (#1072)
* collector_probe: use CAPTURE_* defines for sched/ tracepoints

Install the sched/ tracepoints only if the respective capture
symbolic constant is defined.  Also w/o BPF_SUPPORTS_RAW_TRACEPOINTS,
the `schred_process_fork` tracepoint cannot be used and, hence,
disable it.

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>

* collector_probe: handle catch-all vs. direct syscall tracepoints

The Falco libs BPF driver is designed to catch all syscalls through
the sys_enter and sys_exit tracepoints.  The collector custom
probe attaches to particular system calls only.  In that case,
the tracepoint context depends on the syscall with a syscall-specific
format (structure).

This commit converts the syscall-specific format into the format
sys enter format expected by Falco's BPF driver and filler programs.
Previously, the custom probe copied data from beyond the tracepoint
context. This resulted in offset check failure and a permission denied
when attaching BPF probe (see also notes in the custom probe).

This is the sys enter path.  For the sys exit path, the syscall-specific
exit format (which is identical for direct attached syscalls) needs
to be aligned (wrt. to offsets) with the 'catch-all' sys exit format.

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>

* collector_probe: add USE_COLLECTOR_CUSTOM_PROBES

Introduce a symbolic constant to determine if custom collector
probe or the legacy collector probe (RHEL-7 / "catch-all") is
used.

Note that this is also required to align the syscall tracepoint
formats on the exit event of direct vs. generic syscall exit.

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>

* collector_probe: add socketcall support

For s390x, glibc, Golang use the "socketcall" syscall to multiplex
socket calls (e.g. connect, accept, ...).  This commit provides
the necessary entry/exit path to handle socketcalls.

Respective socketcall handling is also required in the Falco BPF
driver.  For the collector custom probe, additional changes are
required because BPF_SUPPORTS_RAW_TRACEPOINTS is not used.  Also
the socketcall implementation filters socket calls according to
the defined collector syscalls (on enter and exit syscall path).
For the syscall enter, also do pre-filtering in the custom probe
because of direct access to the socketcall arguments.

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>

* collector_probe: correct few typos and style

Co-authored-by: Giles Hutton <ghutton@redhat.com>
Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>

* collector_probe: correct accept4 argument copying

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>

* collector_probe: Fix integer variable size on arguments

Suggested-by: Hendrik Brueckner <brueckner@linux.ibm.com>
Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>

* collector_probe: Re-adding the sched events

Co-authored-by: Hendrik Brueckner <brueckner@linux.ibm.com>
Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>

* collector: rebase for falco version bump

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>

* collector_probe: syscall-based socketcall support

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>

* enable ebpf on s390x

* Bump MODULE_VERSION to 2.4.0-rc2

* fix lint errors in collector_probe.c

* fix lint errors in collector_probe.c

* fix empty socketcall args in ebpf probe on z

* Revert "fix empty socketcall args in ebpf probe on z"

This reverts commit 3e216361a55c97950d7159e7476da28cbd69a271.

* fix empty socketcall args in ebpf probe on z by using new helper function

* fix sys_enter_setgid_args to have same format as /sys/kernel/tracing/events/syscalls/sys_enter_setgid/format

* fixup other syscall structs in collector_probe to match syscall format

* fix lint error in collector_probe.c

* cleanup collector_probe.c

* update prepare-src to copy new collector-probe.h file

* address code review comments

* Remove USE_COLLECTOR_CUSTOM_PROBES

* fix typo

* remove whitespace in collector_probe.h

* fix lint issue

* update falcosecurity-libs submodule

* Update MODULE_VERSION to 2.5.0-rc2

---------

Signed-off-by: Hendrik Brueckner <brueckner@linux.ibm.com>
Signed-off-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>
Co-authored-by: Hendrik Brueckner <brueckner@linux.ibm.com>
Co-authored-by: Giles Hutton <ghutton@redhat.com>
Co-authored-by: Mauro Ezequiel Moltrasio <mmoltras@redhat.com>
Co-authored-by: Kirk Crane <kcrane@redhat.com>