Allow to run Collector without grpc connection (#1203)
First step towards standalone Collector for development. Replace
SignalServiceClient for ProcessSignalHandler with an interface and allow
passing a dummy implementation to dump everything into stdout. Make all
network tracking infrastructure to not depend on the GRPC channel present.

In that case, output everything into stdout. Having this in place, one can
remove --grpc-server argument when building an image, and Collector will not
try to connect anywhere.