Provide the sha256 hash of drivers used
Providing this hash is meant for us to be able to properly identify a
driver. This is a byproduct of us building drivers in 2 separate CI
systems, since we'll eventually unify the systems, I'd expect this
mechanism to be deprecated at some point. However, printing the sha256
to the logs could also prove a useful resource for users if we decide to
provide individual checksums for the drivers in the future.