Use non-docker actionlint commit hook
The actionlint-docker commit hook has some issues when running using
podman from macos, it tries to mount the repository in a way that is not
supported and fails. Because fixing this upstream would be annoying and
installing actionlint locally shouldn't be a big deal, we should start
using the native binary option. This is also preferable to committing
with --no-verify and relaying on people running the check manually.