Implement builder versioning
Versioning for the builder has been changed so that tags and release
branches build their own versions without conflicting with master and
PRs. The builder has also undergone a slight refactor so instead of
using the entire repository as its context, just the builder directory
is used.

We also removed luajit from the list of dependencies, this hasn't been
needed since the removal of chisels on #1290.