Remove luajit dependency
As part of the commit, versioning for the builder has been changed so
that tags and release branch build their own versions without
conflicting with master and PRs. The builder has also undergone a slight
refactor so instead of using the entire repository as its context, just
the builder directory is used.