Use the effective URL for logging when downloading a driver
Loosely related to ROX-19614, when sensor and collector are deployed to
a namespace other than stackrox, the URL logged by collector points to
it trying to download the driver from sensor.stackrox.svc when it should
be using sensor.<namespace>.svc.

This change adds a GetEffectiveURL method to the FileDownloader class,
so we can now print the correct URL the driver will be downloaded from.