Fix master branch retagging full image instead of creating manifests
After merging PR #1373, master started retagging the x86 full image
instead of creating multiarch manifests. This was an unintended
condition that was caused because of a previous simplification that was
no longer true and did not manifest in the PR. This patch makes the
conditions for the full image more strict.