Fix multiarch full image builds on PRs
There's a bug when building full multiarch images on PRs. The bucket
used for pulling drivers is changed in the main CI workflow to use a
staging bucket, but the main CI never pushes drivers for P/Z to this
bucket, so it fails because it find no drivers to add into the image.
Changing to the staging bucket should only be done by the CPaaS
workflow, so an additional condition for this is added.