Use vanilla Falco
This change makes it so collector compiles and runs using upstream Falco
directly. There are a few minor tweaks still required in Falco itself,
but they are things that are either on their way to upstream or should
be real ease to upstream.

List of changes needed for vanilla to be compatible:
- Remove sinsp-wrapper and link directly to sinsp.
- Remove sinsp_filter_check_iface in favor of sinsp_filter_check.
- Use MINIMAL_BUILD instead of MINIMAL_BUILD_WITH_EBPF.
- Remove g_bpf_drop_syscalls.
- eBPF probe is compiled directly instead of via cmake.