Explicitly add procexit event
One side effect of [1] is that procexit is not included by default
any more if ppm_sc is not empty. Which means to continue receiving
procexit we need to explicitly add it into the list of interesting
events. procexit is essential for memory management, as its handlers
perform threadinfo cache clean up.

[1]: https://github.com/falcosecurity/libs/pull/1001