Skip health check for CPaaS workflow
Currently, we use health check to verify Collector has started properly,
but not all the pipelines have the health check enabled in the image.
It's fine, because we want to use it only for testing, so make the
integration tests smarter to handle situations where there is no health
check in the image.