Ignore null events after assert
libsinsp has some number of ASSERT(false) in places that considered to
be impossible to reach, one of them in sinsp::next when doing delayed
removal of fds. This approach is suboptimal, so in 01496ee1a it was
modified to only log the issue and proceed.

There seem to be a negative consequence of those two factors above
combined. While considering an ASSERT impossible to reach, there is a
return right after it, which we now reach if ASSERT doesn't stop the
execution. This returns a current processing status (could be SUCCESS),
yet the event object is not yet passed by reference back to the called,
and SysdigService::GetNext ends up with a success status and null
pointer for an event.

As a workaround, be more vigilant about the events we process.