Use debug for stdout_callback (#1572)
Since integration tests are invoked from the ansible directory, they
pick up configuration from `ansible.cfg`. Currently, it says that the
stdout_callback has to be selective, which means only failed steps are
going to be shown. An annoying consequence is that stdout will not be
exposed even for failing steps (unless handled explicitly), which makes
troubleshooting of some steps harder. Make it debug and verify if this
will produce too much logs.