Configure detailed metrics
Currently, Collector exposes two types of prometheus metrics: regular
aggregates and per-syscall aggregates. The latter one is quite verbose
and produces more than half of the whole output (181 out of 284 lines),
thus it makes sense to make it configurable to be able to disable in the
future.

From one side it would enable us to finally expose Collector prometheus
metrics, at the same we still would be able to use per-syscall metrics
for troubleshooting in development environment.