Split slim builds in local gha runner and remote vm
This separation should make it a little easier to follow the workflows,
since each of them will only have the steps needed for the type of build
taking place. Additional motivation for this is in having better
handling of external PRs, since a PR from a fork will not have the
required credentials to create a remote VM in our GCP projects.