Use event watchers for k8s based integration tests
Listing events is failing for some reason and I've not been able to get
it to reliably output all events for a given namespace. In order to
ensure we retrieve all relevant events, we can spawn a watcher alongside
any namespace we create and have a separate goroutine dumping to a file.