Adjust buffer size instead of cpu-per-core
It turns out making many CPU cores working with a single ringbuf has a
potential deadlock hazard, especially visible on RT kernels. To avoid
even a slightest danger, reduce ringbuf size instead. In the new
scenario if we do not fit into the total ringbuf size, the size of each
individual ringbuf will be adjusted and aligned for readability.