Small fix to libbpf log downgrading
The check for log level needs to happen after we've downgraded them to
trace. In order to prevent looking for the `libbpf:` prefix on every log
line issued by Falco, we first check if we are running in debug mode,
limiting the impact to systems that are actively being debugged.