Run all integration tests
We split the set of integration tests run on PRs due to some of them
taking a long time and some flakiness. Since we did this split, the
flakiness has gone down as well as the time taken to run the tests, so
we might as well always run the full set.