Update Falco bumping instructions
With the removal of the MODULE_VERSION file in #1773 our handling of
the falco libs fork is somewhat outdated, since now there is no module
version to use as a guideline. Instead, I'm proposing that we start
using the upstream version we are based on as part of our branching
strategy.

With this new approach, branches with our changes will be of the form
M.m.p-stackrox, with M.m.p being the version numbers for the upstream
libs tag the branch is based on.