ROX-25477: Exercise integration tests on large nodes (#1763)
A node with large number of CPU cores is an important edge case for
Collector. Run integration tests on such a setup if requested
explicitely or by schedule.

Add a new integration test, which when invoked on a machine with large
number of CPU cores will test the ring buffer adjustment, in particular
that the final adjusted value land in an allowed interval of power of
two.