Provide environment variable alternatives to CLI arguments
This change makes it so environment variables can be used to configure
all arguments that can be provided via CLI parameters. This should in
turn make configuring collector in k8s easier and simplify the command
to be used in our dockerfiles.

In order to prevent as many breaking changes as possible, CLI arguments
will take precedence over environment variables. For TLS certificates a
variable was added for specifying a directory holding the ca.pem,
cert.pem and key.pem files, and three other variables were added to be
able to override the location of each of the files or explicitly setting
all three if needed.

New environment variables added:
- ROX_COLLECTOR_LOG_LEVEL: Sets the log level to be used.
- ROX_COLLECTOR_SCRAPE_INTERVAL: Set the scrapping interval in seconds.
- ROX_COLLECTOR_SCRAPE_DISABLED: Completely disable scrapping.
- ROX_COLLECTOR_GRPC_SERVER: URL to the GRPC server.
- ROX_COLLECTOR_TLS_CERTS: Path to a directory holding TLS certificates.
- ROX_COLLECTOR_TLS_CA: Path to the CA certificate to be used in TLS.
- ROX_COLLECTOR_TLS_CLIENT_CERT: Path to the client certificate to be used in TLS.
- ROX_COLLECTOR_TLS_CLIENT_KEY: Path to the client key to be used in TLS.